tidy = death

+++ tasks/seq/Sadikov_I_SparesMatrixMultiplication/src/ops_seq.cpp
@@ -1,7 +1,5 @@
 #include "seq/Sadikov_I_SparesMatrixMultiplication/include/ops_seq.hpp"
 
-#include <algorithm>
-#include <cstdint>
 #include <random>
 #include <vector>
 
@@ -55,7 +53,7 @@ std::vector<double> sadikov_i_sparse_matrix_multiplication_task_seq::GetRandomMa
   std::random_device dev;
   std::mt19937 gen(dev());
   for (auto i = 0; i < size; ++i) {
-    data[i] = gen() % 500;
+    data[i] = static_cast<double>(gen() % 500);
     if (data[i] > 250.0) {
       data[i] = 0.0;
     }
